+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [Unreleased] - ReleaseDate
 
+- Added method `SrcImageView.set_crop_box_to_fit_dst_size()`.
+- Fixed out-of-bounds error during resize with cropping.
+- Refactored `ImageData`. 
+  - Added methods: `from_vec_u32()`, `from_vec_u8()`, `from_slice_u32()`,
+    `from_slice_u8()`.
+  - Removed methods: `from_buffer()`, `from_pixels()`.
 
 ## [0.2.0] - 2021-08-02